added application received status, moved other constants into constants file
